Moved classes to separate files
Also renamed a ton of variables and swapped out arraylist of dictionary
