+++ WebApi/Triggers/TreeTriggers.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
@@ -9,86 +10,17 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 
-using MySql.Data;
 using MySql.Data.MySqlClient;
 
 namespace ForestGEO.WebApi.Triggers.Tree
 {
-    public class TreeRequest
-    {
-        public string Subquadrat { get; set; }
-        public string Tag { get; set; }
-        public string StemTag { get; set; }
-        public string SpCode { get; set; }
-        public double DBH { get; set; }
-        public string Codes { get; set; }
-        public string Comments { get; set; }
-        public bool WasDead(ArrayList Trees){
-            foreach(TreeStorage tree in Trees){
-                if(tree.Tag == Tag && tree.StemTag == StemTag
-                    && tree.Codes.Contains("dt") ){
-                    return true;
-                }
-            }
-            return false;
-        }
-    }
-    public class TreeStorage {
-        public int TempID { get; set; }
-        public string QuadratName { get; set; }
-        public string Tag { get; set; }
-        public string StemTag { get; set; }
-        public string SpCode { get; set; }
-        public double DBH { get; set; }
-        public string Codes { get; set; }
-        public string HOM { get; set; }
-        public string ExactDate { get; set; }
-        public double x { get; set; }
-        public double y { get; set; }
-        public string PlotID { get; set; }
-        public string CensusID { get; set; }
-        public string Errors { get; set; }
-
-        public TreeStorage MapSQL(MySqlDataReader rdr){
-            TempID = rdr.GetInt32("TempID");
-            QuadratName = rdr["QuadratName"].ToString();
-            Tag = rdr["Tag"].ToString();
-            StemTag = rdr["StemTag"].ToString();
-            SpCode = rdr["SpCode"].ToString();
-            DBH = rdr.GetDouble("DBH");
-            Codes = rdr["Codes"].ToString();
-            HOM = rdr["HOM"].ToString();
-            ExactDate = rdr["ExactDate"].ToString();
-            x = rdr.GetDouble("x");
-            y = rdr.GetDouble("y");
-            PlotID = rdr["PlotID"].ToString();
-            CensusID = rdr["CensusID"].ToString();
-            Errors = rdr["Errors"].ToString();
-            return this;
-        }
-    }
-    public class TreeResponse
-    {
-        public string Subquadrat { get; set; }
-        public string Tag { get; set; }
-        public string StemTag { get; set; }
-        public int ErrorCode { get; set; }
-        public string Error { get; set; }
-        public TreeResponse (TreeRequest tree, int ecode, string error){
-            Subquadrat = tree.Subquadrat;
-            Tag = tree.Tag;
-            StemTag = tree.StemTag;
-            ErrorCode = ecode;
-            Error = error;
-        }
-    }
-    public static class TreeDataMock
+    public static class ProcessTrees
     {
         private static string connStr = System.Environment.GetEnvironmentVariable("MySQLConnection", EnvironmentVariableTarget.Process);
         private static MySqlConnection conn = new MySqlConnection(connStr);
 
-        private static ArrayList QueryTreeDB(string sql){
-            var LoadingResponse = new ArrayList();
+        private static Dictionary<(string,string),TreeStorage> QueryTreeDB(string sql){
+            var LoadingResponse = new Dictionary<(string,string),TreeStorage>();
             try
             {
                 Console.WriteLine("Connecting to MySQL...");
@@ -98,7 +30,11 @@ namespace ForestGEO.WebApi.Triggers.Tree
                 MySqlDataReader rdr = cmd.ExecuteReader();
                 
                 while (rdr.Read())
-                    LoadingResponse.Add(new TreeStorage().MapSQL(rdr));
+                {
+                    var tempTreeStorage = new TreeStorage().MapSQL(rdr);
+                    LoadingResponse.Add((tempTreeStorage.Tag,tempTreeStorage.StemTag), tempTreeStorage);
+                }
+                    
                 rdr.Close();
             }
             catch (Exception ex)
@@ -119,7 +55,7 @@ namespace ForestGEO.WebApi.Triggers.Tree
         {
             log.LogInformation("C# HTTP trigger function processed a request.");
 
-            ArrayList LoadingResponse = QueryTreeDB("select TempID, QuadratName, Tag, StemTag, Mnemonic as SpCode, DBH, Codes, HOM, ExactDate, x, y, PlotID, CensusID, Errors from tempnewplants");
+            Dictionary<(string,string),TreeStorage> LoadingResponse = QueryTreeDB("select TempID, QuadratName, Tag, StemTag, Mnemonic as SpCode, DBH, Codes, HOM, ExactDate, x, y, PlotID, CensusID, Errors from tempnewplants");
 
             return new OkObjectResult(JsonConvert.SerializeObject(LoadingResponse));
         }
@@ -132,18 +68,20 @@ namespace ForestGEO.WebApi.Triggers.Tree
             log.LogInformation("C# HTTP trigger function processed a request.");
 
             string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
-            var data = JsonConvert.DeserializeObject<TreeRequest[]>(requestBody);
+            var data = JsonConvert.DeserializeObject<Tree[]>(requestBody);
 
-            ArrayList LoadingResponse = QueryTreeDB("select TempID, QuadratName, Tag, StemTag, Mnemonic as SpCode, DBH, Codes, HOM, ExactDate, x, y, PlotID, CensusID, Errors from tempnewplants");
+            Dictionary<(string,string),TreeStorage> TreeRecords = QueryTreeDB("select TempID, QuadratName, Tag, StemTag, Mnemonic as SpCode, DBH, Codes, HOM, ExactDate, x, y, PlotID, CensusID, Errors from tempnewplants");
 
-            var TreeResponse = new ArrayList();
+            var ResponseBuilder = new ArrayList();
             foreach(var Tree in data){
-                if(Tree.Codes.Contains("at") && Tree.WasDead(LoadingResponse)){
-                    TreeResponse.Add(new TreeResponse(Tree, 1, "This tree was dead in a previous census."));
+                if(TreeRecords.ContainsKey((Tree.Tag, Tree.StemTag)) && 
+                    Tree.IsAlive() && TreeRecords[(Tree.Tag, Tree.StemTag)].IsDead())
+                {
+                    ResponseBuilder.Add(new TreeResponse(Tree, 1, "This tree was dead in a previous census."));
                 }
             }
 
-            return new OkObjectResult(JsonConvert.SerializeObject(TreeResponse));
+            return new OkObjectResult(JsonConvert.SerializeObject(ResponseBuilder));
         }
     }
 }
